Cleanup Auxiliary Linux tests
This PR merges the "extra" Linux tests into a single job. This reduces
the number of jobs we have to run as part of the CI. We also explictly
install libc and libgcc (to make it clear these are the only extra dev
dependancies).

I also added building/linking of the x32 target. As noted in #464, this
will not run in the current github actions runner due to an incompatible
kernel config.

Signed-off-by: Joe Richey <joerichey@google.com>